MODHAADM-71 separate test suites for exclude/include Harvester
  Adding profile for running tests against installed and running Harvester, activate with `-PharvesterTests`
  Default is to run unit tests that doesn't require Harvester
