+++ src/test/java/org/folio/harvesteradmin/test/HarvesterIntegrationTestSuite.java
@@ -1,6 +1,5 @@
 package org.folio.harvesteradmin.test;
 
-import static io.restassured.RestAssured.given;
 import static org.folio.harvesteradmin.legacydata.statics.ApiPaths.THIS_HARVESTABLES_PATH;
 import static org.folio.harvesteradmin.legacydata.statics.ApiPaths.THIS_STEPS_PATH;
 import static org.folio.harvesteradmin.legacydata.statics.ApiPaths.THIS_STORAGES_PATH;
@@ -17,18 +16,10 @@ import static org.folio.harvesteradmin.test.sampleData.Samples.SAMPLE_STEP;
 import static org.folio.harvesteradmin.test.sampleData.Samples.SAMPLE_STEP_2;
 import static org.folio.harvesteradmin.test.sampleData.Samples.SAMPLE_STEP_2_ID;
 import static org.folio.harvesteradmin.test.sampleData.Samples.SAMPLE_STEP_ID;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertTrue;
 
 import io.restassured.RestAssured;
-import io.restassured.builder.RequestSpecBuilder;
-import io.restassured.config.HttpClientConfig;
-import io.restassured.http.ContentType;
-import io.restassured.http.Header;
-import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import io.restassured.specification.RequestSpecification;
 import io.vertx.core.DeploymentOptions;
 import io.vertx.core.Vertx;
 import io.vertx.core.json.JsonArray;
@@ -39,7 +30,6 @@ import io.vertx.ext.unit.junit.VertxUnitRunner;
 import org.folio.harvesteradmin.MainVerticle;
 import org.folio.harvesteradmin.test.fakestorage.FakeFolioApis;
 import org.folio.harvesteradmin.test.utils.NetworkUtils;
-import org.folio.okapi.common.XOkapiHeaders;
 import org.folio.tlib.postgres.testing.TenantPgPoolContainer;
 import org.junit.*;
 import org.junit.rules.TestName;
@@ -48,26 +38,14 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.testcontainers.containers.PostgreSQLContainer;
 
-import java.time.LocalDateTime;
-import java.time.temporal.ChronoUnit;
-import java.util.UUID;
-
 @RunWith( VertxUnitRunner.class )
-public class HarvesterAdminTestSuite {
+public class HarvesterIntegrationTestSuite {
   private static final Logger logger = LoggerFactory.getLogger( "HarvesterAdminTestSuite" );
 
-  static final String TENANT = "mha_test";
   static Vertx vertx;
   public static int PORT_HARVESTER_ADMIN = NetworkUtils.nextFreePort();
-  public static final int PORT_OKAPI = 9031;
-  public static final Header CONTENT_TYPE_JSON = new Header("Content-Type", "application/json");
-  public static final Header CONTENT_TYPE_XML = new Header("Content-Type", "application/xml");
-  public static final Header OKAPI_TENANT = new Header (XOkapiHeaders.TENANT, TENANT);
-  public static final Header OKAPI_URL = new Header (XOkapiHeaders.URL, "http://localhost:" + PORT_OKAPI);
-  public static final Header OKAPI_TOKEN = new Header(XOkapiHeaders.TOKEN,"eyJhbGciOiJIUzUxMiJ9eyJzdWIiOiJhZG1pbiIsInVzZXJfaWQiOiI3OWZmMmE4Yi1kOWMzLTViMzktYWQ0YS0wYTg0MDI1YWIwODUiLCJ0ZW5hbnQiOiJ0ZXN0X3RlbmFudCJ9BShwfHcNClt5ZXJ8ImQTMQtAM1sQEnhsfWNmXGsYVDpuaDN3RVQ9");
 
-
-  public HarvesterAdminTestSuite() {}
+  public HarvesterIntegrationTestSuite() {}
 
   @ClassRule
   public static PostgreSQLContainer<?> postgresSQLContainer = TenantPgPoolContainer.create();
@@ -84,7 +62,6 @@ public class HarvesterAdminTestSuite {
     // Register the testContext exception handler to catch assertThat
     vertx.exceptionHandler(testContext.exceptionHandler());
     RestAssured.port = PORT_HARVESTER_ADMIN;
-
     System.setProperty("port", String.valueOf(PORT_HARVESTER_ADMIN));
     vertx.deployVerticle(
         MainVerticle.class.getName(), new DeploymentOptions())
@@ -111,39 +88,6 @@ public class HarvesterAdminTestSuite {
     deleteRecordsByIdPrefix(THIS_STEPS_PATH, "transformationSteps");
   }
 
-  void initModuleDatabase () {
-    tenantOp(TENANT, new JsonObject().put("module_to", "mod-harvester-admin-0.0.0"), null);
-  }
-
-  void tenantOp(String tenant, JsonObject tenantAttributes, String expectedError) {
-    ExtractableResponse<Response> response = RestAssured.given()
-            .header(XOkapiHeaders.TENANT, tenant)
-            .contentType(ContentType.JSON)
-            .body(tenantAttributes.encode())
-            .post("/_/tenant")
-            .then()
-            .extract();
-
-    if (response.statusCode() == 204) {
-      return;
-    }
-    assertThat(response.statusCode(), is(201));
-    String location = response.header("Location");
-    JsonObject tenantJob = new JsonObject(response.asString());
-    assertThat(location, is("/_/tenant/" + tenantJob.getString("id")));
-
-    RestAssured.given()
-            .header(XOkapiHeaders.TENANT, tenant)
-            .get(location + "?wait=10000")
-            .then().statusCode(200)
-            .body("complete", is(true))
-            .body("error", is(expectedError));
-
-    RestAssured.given()
-            .header(XOkapiHeaders.TENANT, tenant)
-            .delete(location)
-            .then().statusCode(204);
-  }
 
 
   private void deleteRecordsByIdPrefix(String path, String recordsArrayProperty) {
@@ -681,142 +625,6 @@ public class HarvesterAdminTestSuite {
         Response resp3 = postConfigRecord(harvestable, THIS_HARVESTABLES_PATH, 422);
         assertTrue(resp3.body().asPrettyString().contains("name 'Test harvest job' exists already"));
 
-  }
-
-  @Test
-  public void canGetPreviousJob () {
-    initModuleDatabase();
-    Response response = given().port(PORT_HARVESTER_ADMIN)
-            .header(OKAPI_TENANT)
-            .get("harvester-admin/previous-jobs")
-        .then()
-        .log().ifValidationFails().statusCode(200).extract().response();
-    logger.info("canGetPreviousJobs response: " + response.asPrettyString());
-  }
-
-  @Test
-  public void willPurgeAgedJobLogsUsingDefaultThreshold () {
-    initModuleDatabase();
-
-    Response response = given().port(PORT_HARVESTER_ADMIN)
-            .header(OKAPI_TENANT)
-            .get("harvester-admin/previous-jobs")
-            .then()
-            .log().ifValidationFails().statusCode(200).extract().response();
-    logger.info("will purge jobs response: " + response.asPrettyString());
-
-    given().port(PORT_HARVESTER_ADMIN)
-            .header(OKAPI_TENANT)
-            .get("harvester-admin/previous-jobs")
-            .then()
-            .log().ifValidationFails().statusCode(200).extract().response();
-
-    LocalDateTime now = LocalDateTime.now();
-    final LocalDateTime agedJobStartedTime = now.minusMonths(3).minusDays(1).truncatedTo(ChronoUnit.SECONDS);
-    final LocalDateTime agedJobFinishedTime = agedJobStartedTime.plusMinutes(2);
-    final LocalDateTime newerJobStartedTime = now.minusMonths(2).truncatedTo(ChronoUnit.SECONDS);
-    final LocalDateTime newerJobFinishedTime = newerJobStartedTime.plusMinutes(3);
-
-    JsonObject transformation = BASE_TRANSFORMATION_JSON;
-    String transformationId = transformation.getString("id");
-    postConfigRecord(BASE_TRANSFORMATION_JSON, THIS_TRANSFORMATIONS_PATH, 201);
-
-    JsonObject agedJobJson =
-            new JsonObject(
-                    "    {\n" +
-                    "      \"id\" : \"" + UUID.randomUUID() + "\",\n" +
-                    "      \"name\" : \"fake job log\",\n" +
-                    "      \"harvestableId\" : 672813240090200,\n" +
-                    "      \"type\" : \"xmlBulk\",\n" +
-                    "      \"url\" : \"http://fileserver/xml/\",\n" +
-                    "      \"allowErrors\" : true,\n" +
-                    "      \"transformation\" : \""+ transformationId +"\",\n" +
-                    "      \"storage\" : \"Batch Upsert Inventory\",\n" +
-                    "      \"status\" : \"OK\",\n" +
-                    "      \"started\" : \""+agedJobStartedTime+"\",\n" +
-                    "      \"finished\" : \""+agedJobFinishedTime+"\",\n" +
-                    "      \"amountHarvested\" : 5,\n" +
-                    "      \"message\" : \"  Instances_processed/loaded/deletions(signals)/failed:__5___5___0(0)___0_ Holdings_records_processed/loaded/deleted/failed:__13___13___0___0_ Items_processed/loaded/deleted/failed:__4___4___0___0_ Source_records_processed/loaded/deleted/failed:__0___0___0___0_\"\n" +
-                    "    }\n");
-
-    JsonObject newerJobJson =
-            new JsonObject(
-                    "    {\n" +
-                            "      \"id\" : \"" + UUID.randomUUID() + "\",\n" +
-                            "      \"name\" : \"fake job log\",\n" +
-                            "      \"harvestableId\" : 672813240090200,\n" +
-                            "      \"type\" : \"xmlBulk\",\n" +
-                            "      \"url\" : \"http://fileserver/xml/\",\n" +
-                            "      \"allowErrors\" : true,\n" +
-                            "      \"transformation\" : \"" + transformationId + "\",\n" +
-                            "      \"storage\" : \"Batch Upsert Inventory\",\n" +
-                            "      \"status\" : \"OK\",\n" +
-                            "      \"started\" : \""+newerJobStartedTime+"\",\n" +
-                            "      \"finished\" : \""+newerJobFinishedTime+"\",\n" +
-                            "      \"amountHarvested\" : 3,\n" +
-                            "      \"message\" : \"  Instances_processed/loaded/deletions(signals)/failed:__3___3___0(0)___0_ Holdings_records_processed/loaded/deleted/failed:__8___8___0___0_ Items_processed/loaded/deleted/failed:__2___2___0___0_ Source_records_processed/loaded/deleted/failed:__0___0___0___0_\"\n" +
-                            "    }\n");
-
-    given().port(PORT_HARVESTER_ADMIN).header(OKAPI_TENANT)
-            .body(agedJobJson.encode())
-            .contentType(ContentType.JSON)
-            .post("harvester-admin/previous-jobs")
-            .then()
-            .log().ifValidationFails().statusCode(201).extract().response();
-
-    given().port(PORT_HARVESTER_ADMIN).header(OKAPI_TENANT)
-            .body(newerJobJson.encode())
-            .contentType(ContentType.JSON)
-            .post("harvester-admin/previous-jobs")
-            .then()
-            .log().ifValidationFails().statusCode(201).extract().response();
-
-    RestAssured
-            .given()
-            .port(PORT_HARVESTER_ADMIN)
-            .header(OKAPI_TENANT)
-            .contentType(ContentType.JSON)
-            .get("harvester-admin/previous-jobs")
-            .then().statusCode(200)
-            .body("totalRecords",  is(2));
-
-    final RequestSpecification timeoutConfig = timeoutConfig(10000);
-
-    given()
-            .port(PORT_OKAPI)
-            .header(OKAPI_TENANT)
-            .header(OKAPI_URL)
-            .header(OKAPI_TOKEN)
-            .contentType(ContentType.JSON)
-            .header(XOkapiHeaders.REQUEST_ID, "purge-aged-logs")
-            .spec(timeoutConfig)
-            .when().post("/harvester-admin/purge-aged-logs")
-            .then().log().ifValidationFails().statusCode(204)
-            .extract().response();
-
-    RestAssured
-            .given()
-            .port(PORT_HARVESTER_ADMIN)
-            .header(OKAPI_TENANT)
-            .contentType(ContentType.JSON)
-            .get("harvester-admin/previous-jobs")
-            .then().statusCode(200)
-            .body("totalRecords",  is(1));
-
-  }
-
-  @Test
-  public void willPurgeAgedJobLogsUsingConfiguredThreshold () {
-
-  }
-
-  public static RequestSpecification timeoutConfig(int timeOutInMilliseconds) {
-    return new RequestSpecBuilder()
-            .setConfig(RestAssured.config()
-                    .httpClient(HttpClientConfig.httpClientConfig()
-                            .setParam("http.connection.timeout", timeOutInMilliseconds)
-                            .setParam("http.socket.timeout", timeOutInMilliseconds)))
-            .build();
   }
 
   @Test
