+++ src/test/java/org/folio/harvesteradmin/test/Statics.java
@@ -1,0 +1,70 @@
+package org.folio.harvesteradmin.test;
+
+import io.restassured.RestAssured;
+import io.restassured.builder.RequestSpecBuilder;
+import io.restassured.config.HttpClientConfig;
+import io.restassured.http.ContentType;
+import io.restassured.http.Header;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import io.restassured.specification.RequestSpecification;
+import io.vertx.core.json.JsonObject;
+import org.folio.okapi.common.XOkapiHeaders;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.is;
+
+public class Statics {
+
+    static final String TENANT = "mha_test";
+    public static final int PORT_OKAPI = 9031;
+    public static final Header CONTENT_TYPE_JSON = new Header("Content-Type", "application/json");
+    public static final Header CONTENT_TYPE_XML = new Header("Content-Type", "application/xml");
+    public static final Header OKAPI_TENANT = new Header (XOkapiHeaders.TENANT, TENANT);
+    public static final Header OKAPI_URL = new Header (XOkapiHeaders.URL, "http://localhost:" + PORT_OKAPI);
+    public static final Header OKAPI_TOKEN = new Header(XOkapiHeaders.TOKEN,"eyJhbGciOiJIUzUxMiJ9eyJzdWIiOiJhZG1pbiIsInVzZXJfaWQiOiI3OWZmMmE4Yi1kOWMzLTViMzktYWQ0YS0wYTg0MDI1YWIwODUiLCJ0ZW5hbnQiOiJ0ZXN0X3RlbmFudCJ9BShwfHcNClt5ZXJ8ImQTMQtAM1sQEnhsfWNmXGsYVDpuaDN3RVQ9");
+
+    static void initModuleDatabase () {
+        tenantOp(Statics.TENANT, new JsonObject().put("module_to", "mod-harvester-admin-0.0.0"), null);
+    }
+
+    static void tenantOp(String tenant, JsonObject tenantAttributes, String expectedError) {
+        ExtractableResponse<Response> response = RestAssured.given()
+                .header(XOkapiHeaders.TENANT, tenant)
+                .contentType(ContentType.JSON)
+                .body(tenantAttributes.encode())
+                .post("/_/tenant")
+                .then()
+                .extract();
+
+        if (response.statusCode() == 204) {
+            return;
+        }
+        assertThat(response.statusCode(), is(201));
+        String location = response.header("Location");
+        JsonObject tenantJob = new JsonObject(response.asString());
+        assertThat(location, is("/_/tenant/" + tenantJob.getString("id")));
+
+        RestAssured.given()
+                .header(XOkapiHeaders.TENANT, tenant)
+                .get(location + "?wait=10000")
+                .then().statusCode(200)
+                .body("complete", is(true))
+                .body("error", is(expectedError));
+
+        RestAssured.given()
+                .header(XOkapiHeaders.TENANT, tenant)
+                .delete(location)
+                .then().statusCode(204);
+    }
+
+    public static RequestSpecification timeoutConfig(int timeOutInMilliseconds) {
+        return new RequestSpecBuilder()
+                .setConfig(RestAssured.config()
+                        .httpClient(HttpClientConfig.httpClientConfig()
+                                .setParam("http.connection.timeout", timeOutInMilliseconds)
+                                .setParam("http.socket.timeout", timeOutInMilliseconds)))
+                .build();
+    }
+
+}
